+++ src/components/header/logo/logo.tsx
@@ -1,7 +1,6 @@
 import { Link } from 'react-router-dom';
 import { AppRoute } from '../../../const';
 import classNames from 'classnames';
-import { memo } from 'react';
 
 type LogoProps = {
   isMainPage: boolean;
@@ -21,4 +20,4 @@ function Logo({isMainPage}:LogoProps): JSX.Element {
   );
 }
 
-export default memo(Logo);
+export default Logo;
