+++ src/components/layout/use-layout.ts
@@ -1,0 +1,32 @@
+import { useLocation } from 'react-router-dom';
+import { AppRoute } from '../../const';
+import { useAppSelector } from '../../hooks/redux';
+import { getIsFavoritesOffers } from '../../store/offers-data/selectors';
+
+
+export function useLayout(){
+  const {pathname} = useLocation();
+  const isFavoritesOffers = useAppSelector(getIsFavoritesOffers);
+
+  let isMainPage = false;
+  let isLoginPage = false;
+  let rootClassName = '';
+
+  switch (pathname) {
+    case AppRoute.Main:
+      isMainPage = true;
+      rootClassName = ' page--gray page--main';
+      break;
+    case AppRoute.Login:
+      isLoginPage = true;
+      rootClassName = ' page--gray page--login';
+      break;
+    case AppRoute.Favorites:
+      if (!isFavoritesOffers) {
+        rootClassName = ' page--favorites-empty';
+      }
+      break;
+  }
+
+  return {isMainPage, isLoginPage, rootClassName};
+}
