+++ src/components/main-screen/offer-card-list/offer-card-list.tsx
@@ -1,4 +1,3 @@
-//import { useState } from 'react';
 import { useOffersSort } from '../../../hooks/use-offers-sort';
 import { Offers } from '../../../types/offers';
 import OfferCard from '../../common/offer-card/offer-card';
@@ -11,13 +10,15 @@ function OfferCardList({offers}: OfferCardListProps) {
   const sortedOffers = useOffersSort(offers);
 
   return (
-    sortedOffers.map((offer) => (
-      <OfferCard
-        offer={offer}
-        classNamePrefix='cities'
-        key={offer.id}
-      />
-    ))
+    <div className="cities__places-list places__list tabs__content">
+      {sortedOffers.map((offer) => (
+        <OfferCard
+          offer={offer}
+          classNamePrefix='cities'
+          key={offer.id}
+        />
+      ))}
+    </div>
   );
 }
