+++ src/components/offer-screen/reviews-list/reviews-list.tsx
@@ -1,7 +1,5 @@
 import { Reviews } from '../../../types/reviews';
-import { formatDate } from '../../../utils/format-date/format-date';
-import StarsRating from '../../common/stars-rating/stars-rating';
-import UserInfo from '../../common/user-info/user-info';
+import OfferReview from '../review/review';
 
 type OfferReviewsListProps = {
   reviews: Reviews;
@@ -13,23 +11,7 @@ function ReviewsList({reviews}: OfferReviewsListProps): JSX.Element {
       <h2 className="reviews__title">Reviews · <span className="reviews__amount">{reviews.length}</span></h2>
       <ul className="reviews__list">
         {
-          reviews.map((review) => {
-            const {id, user, rating, comment, date} = review;
-            const commentDate = formatDate(date);
-
-            return (
-              <li className="reviews__item" key={id}>
-                <UserInfo user={user} variant='review'/>
-                <div className="reviews__info">
-                  <StarsRating rating={rating} classNamePrefix='reviews'/>
-                  <p className="reviews__text">
-                    {comment}
-                  </p>
-                  <time className="reviews__time" dateTime={date}>{commentDate}</time>
-                </div>
-              </li>
-            );
-          })
+          reviews.map((review) => (<OfferReview review={review} key={review.id}/>))
         }
       </ul>
     </>
